requirements.txt
alembic==1.0.7
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==18.2.0
black==18.9b0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
dash==0.39.0
dash-core-components==0.44.0
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
decorator==4.4.0
dominate==2.3.5
entrypoints==0.3
flake8==3.7.7
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.8
ipython-genutils==0.2.0
itsdangerous==1.1.0
jupyter-core==4.4.0
Keras-Preprocessing==1.0.9
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
mkl
more-itertools==5.0.0
nbformat==4.4.0
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.2
pep8==1.7.1
pep8-naming==0.8.2
plotly==3.7.1
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==3.0.0
pyflakes==2.1.1
pyOpenSSL==19.0.0
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
scikit-learn==0.20.0
scipy==1.2.1
six==1.12.0
snowballstemmer==1.2.1
toml==0.10.0
torch==1.0.1.post2
traitlets==4.3.2
tweepy==3.7.0
urllib3==1.24.1
visitor==0.1.3
waitress==1.2.1
Werkzeug==0.15.1
